# Refactor Django Views Tests to use setUpTestData for Performance Improvement