@Itz-Agasta 2025-07-28
hi @Lorygold  I have refactored 2 test files to use `setUpTestData` instead of `setUp` for database object creation:
- test_views.py
- test_alert_googlechat.py

All tests are passing and performance has improved.

Could you please review these changes? If everything looks good, I will continue refactoring the remaining test files that still use setUp. Thank you!