@Lorygold 2025-07-30
The related issue of this PR has been opened to refactor just the Views test (other issues and PRs have been opened in order to refactor other methods), so the changes to `test_alerter_*` are not required in this task